refactor: removed internal barrel imports
