--- src/core/canvas-app/index.js
@@ -1,3 +1,0 @@
-export * from "./CanvasApp.js";
-export * from "./CanvasAppEvent.js";
-export * from "./CanvasAppPauseEvent.js";
