+++ src/core/index.js
@@ -1,1 +1,3 @@
-export * from "./canvas-app/index.js";
+export * from "./canvas-app/CanvasApp.js";
+export * from "./canvas-app/CanvasAppEvent.js";
+export * from "./canvas-app/CanvasAppPauseEvent.js";
