+++ src/events/index.js
@@ -1,11 +1,11 @@
 export * from "./CanvasEvents.js"
 export * from "./EventEmitter.js"
 
-// --[ keyboard ]---------------------------------------------------------------
+// MARK: - keyboard
 export * from "./keyboard/Keyboard.js"
 export * from "./keyboard/KeyboardEvent.js"
 
-// --[ mouse ]------------------------------------------------------------------
+// MARK: - mouse
 export * from "./mouse/MouseButton.js"
 export * from "./mouse/MouseEventProcessor.js"
-export * from "./mouse/MouseEvents.js"
+export * from "./mouse/MouseEvent.js"
